[REWORK] Restructuration of the project to make everything more readable
